Refactor: move @ParameterizedClass lifecycle to dedicated ParameterizedQuarkusExtensionTest subclass
